# docs/swagger.yaml
basePath: /api/v1
definitions:
  Author:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  AuthorInput:
    properties:
      name:
        minLength: 2
        type: string
    required:
    - name
    type: object
  Book:
    properties:
      author:
        $ref: '#/definitions/Author'
      author_id:
        type: string
      available_stock:
        type: integer
      categories:
        items:
          $ref: '#/definitions/domain.Category'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      discount_percentage:
        type: number
      id:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      isbn:
        type: string
      name:
        type: string
      price:
        type: number
      publisher:
        $ref: '#/definitions/Publisher'
      publisher_id:
        type: string
      updated_at:
        type: string
    type: object
...
  Publisher:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      legal_name:
        type: string
      mobile:
        type: string
      state:
        type: string
      trading_name:
        type: string
      updated_at:
        type: string
      zipcode:
        type: string
    type: object
...
  domain.BookFilter:
    properties:
      authorIDs:
        items:
          type: string
        type: array
      categoryIDs:
        items:
          type: string
        type: array
      ids:
        items:
          type: string
        type: array
      isActive:
        type: boolean
      maxPrice:
        format: float64
        type: number
      minPrice:
        format: float64
        type: number
      minStock:
        type: integer
      publisherIDs:
        items:
          type: string
        type: array
      search:
        description: name / author / isbn
        type: string
    type: object
  domain.BookInput:
    properties:
      author_id:
        type: string
      author_name:
        type: string
      available_stock:
        type: integer
      category_ids:
        items:
          type: string
        type: array
      description:
        type: string
      discount_percentage:
        type: number
      image_url:
        type: string
      is_active:
        type: boolean
      isbn:
        type: string
      name:
        type: string
      price:
        type: number
      publisher_id:
        type: string
    required:
    - author_name
    - name
    - publisher_id
    type: object
  domain.BookSearchResult:
    properties:
      items:
        items:
          $ref: '#/definitions/Book'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  domain.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: kolimirakshitha@gmail.com
    name: Rakshitha
  description: Online Bookstore backend (BookNest)
  license:
    name: MIT
  termsOfService: http://booknest.com/terms/
  title: BookNest API
  version: "1.0"
paths:
...
  /authors:
    get:
      description: Lists authors
      parameters:
      - description: Result limit
        in: query
        name: limit
        type: integer
      - description: Result offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Author'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List authors
      tags:
      - Authors
    post:
      consumes:
      - application/json
      description: Creates a new author (admin only)
      parameters:
      - description: Author input
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AuthorInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Author'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create author
      tags:
      - Authors
  /authors/{id}:
    get:
      description: Fetches a single author by its ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Author'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get author by ID
      tags:
      - Authors
  /books:
    get:
      description: Returns a list of books
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Book'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List books
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Creates a new book (admin only)
      parameters:
      - description: Book input
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.BookInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Book'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create book
      tags:
      - Books
  /books/{id}:
    get:
      description: Fetches a single book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Book'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get book by ID
      tags:
      - Books
  /books/filter:
    post:
      consumes:
      - application/json
      description: Filters books by criteria and pagination
      parameters:
      - description: Search by name, author, or ISBN
        in: query
        name: search
        type: string
      - description: Result limit
        in: query
        name: limit
        type: integer
      - description: Result offset
        in: query
        name: offset
        type: integer
      - description: Book filter payload
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.BookFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.BookSearchResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Filter books
      tags:
      - Books
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"